+++ sen2cli/inference/create.py
@@ -5,7 +5,7 @@ from jsonapi_client import Session
 from jsonapi_client.exceptions import DocumentError
 from oauthlib.oauth2 import OAuth2Token
 
-from utils import dict_from_resource
+from ..utils import dict_from_resource
 from .util import DEFAULT_COLUMNS, INFERENCE_SCHEMA
 from ..env import API_BASE_URL
 from ..session.oauth_util import get_user_info
@@ -20,6 +20,7 @@ def create_inference(token: OAuth2Token,
                      temp_range_end: datetime,
                      spatial_subset: str,
                      description: str,
+                     dry_run: bool = False,
                      ) -> dict:
 
   user_info = get_user_info(token)
@@ -52,7 +53,10 @@ def create_inference(token: OAuth2Token,
       inference.knowledgebase = knowldegebase_id
       inference.factbase = factbase_id
       logger.debug(inference.json)
-      inference.commit()
+      if not dry_run:
+        inference.commit()
+      else:
+        logger.info("Dry run. Skipping commit")
       return dict_from_resource(inference, DEFAULT_COLUMNS)
     except DocumentError as e:
       logger.error(f"Could not create inference. Reason: {e}", exc_info=True)